refactor(llmrails): fix package ownership boundaries
